Reapply changes to AttachLayerPopupButton after merge.

+++ extensions/map-layers/src/ui/widget/AttachLayerPopupButton.tsx
@@ -197,7 +197,7 @@ function AttachLayerPanel({ isOverlay, onLayerAttached }: AttachLayerPanelProps)
   const handleYesConfirmation = React.useCallback(async (source: MapLayerSource) => {
 
     const layerName = source.name;
-    if (!!iTwinId && !!iModelId) {
+    if (!!iTwinId && iTwinId !== Guid.empty && !!iModelId) {
       if (await MapLayerSettingsService.deleteSharedSettings(source, iTwinId, iModelId)) {
         const msg = MapLayersUiItemsProvider.localization.getLocalizedString("mapLayers:CustomAttach.RemoveLayerDefSuccess", { layerName });
         IModelApp.notifications.outputMessage(new NotifyMessageDetails(OutputMessagePriority.Info, msg));
